move files to pkg
