+++ pkg/adapter/store.go
