+++ pkg/p2p/gossip.go
