+++ pkg/signer/signer.go
